@@ -11,7 +11,7 @@ public class CommunicationUtils {
     /** Connection Timeout default, milliseconds */
     public static final int CONNECTION_TIMEOUT  = 20000;   // 20 seconds
     /** Default Server Url */
-    public static final String SERVER_URL         = "http://192.168.0.100";
+    public static final String SERVER_URL         = "http://enzeddpc";
     /** Default Server Port */
     public static final int SERVER_PORT         = 8000;
 }
